requirements.txt
Flask
flask-restful
PyJWT
python-dotenv
psycopg2-binary
redis
